Add git-attributes
Git Attributes make it so that git-based applications
like GitHub can get a better understanding of what a
repository actually contains.

Thus in this case, as we are actually writing a markdown
document, source code formats should be linguist-vendored
to not count towards the repository programming language given
on GitHub.