Say hello to build.py
build.py will replace the Makefile with a python solution which neatly integrates all the required
tasks. Main improvements over the current build system are language support and the introduction
of a build/ dir in which everything is created instead of placing stuff all over the place in the
source tree.