Minor readme adjustments
The readme file for OParl 1.0 was not particalarly helpful
for people from outside the project to begin with. To change
that now would - in a way - break semver rules. Thus the
only change is to hint on the new building strategy without
explaining what all that entails.

However, we should probably consider keeping a more permanent
'How to build this' record somewhere for people who may want
to dig deep into OParls' past someday. Whoever that might be.